lobby system finished | player scyn working

+++ Game-FJNN/src/client/GameCEM.java
@@ -2,6 +2,7 @@ package client;
 
 import data.events.client.ToClientMessageEvent;
 import data.events.client.ToClientMessageEventListener;
+import events.DroneTargetEvent;
 import events.entity.EntityPathEvent;
 import events.entity.EntityStatusEvent;
 import events.entity.PlayerMoveEvent;
@@ -10,7 +11,8 @@ import events.inventory.ItemRemoveEvent;
 import events.inventory.ItemSetEvent;
 import events.map.MapBlockAddEvent;
 import events.map.MapBlockDeleteEvent;
-import game.entity.player.Player;
+import game.entity.player.PlayerContext;
+import game.entity.player.playerDrone.DroneHost;
 import game.entity.type.interfaces.PathUser;
 import game.map.Map;
 
@@ -25,19 +27,23 @@ public class GameCEM implements ToClientMessageEventListener{
 	}
 	
 	@Override
-	public void messageFromServer(ToClientMessageEvent event) {
-		if(event.getMessage().getId() != GameCPM.DataPackage_PlayerMoved)System.out.println("recieved incoming Data: " + event.getMessage().getId());		
-		
+	public void messageFromServer(ToClientMessageEvent event) {		
 		try{
 			if(event.getMessage().getId() == GameCPM.DataPackage_PlayerMoved) {
 				PlayerMoveEvent playerMove = gameCM.getClientPackageManager().readPlayerMoveMessage(event.getMessage());
-				
-				Player p = (Player)playerMove.getEntity();
+				if(playerMove == null) {
+					event.setActive(false);
+					return;
+				}				
+				PlayerContext p = (PlayerContext)playerMove.getEntity();
 				p.setLocation(playerMove.getCurrentPixelLocation().getX(), playerMove.getCurrentPixelLocation().getY());
 				p.setVelocity(playerMove.getVelocity());
 				event.setActive(false);
 			}else if(event.getMessage().getId() == GameCPM.DataPackage_EntityPath) {
 				EntityPathEvent entityPath = gameCM.getClientPackageManager().readEntityPathMessage(event.getMessage());
+				if(entityPath == null) {
+					return;
+				}
 				
 				entityPath.getEntity().setLocation(entityPath.getCurrentPixelLocation().getX(), entityPath.getCurrentPixelLocation().getY());
 				PathUser pathUser = (PathUser)entityPath.getEntity();
@@ -49,34 +55,67 @@ public class GameCEM implements ToClientMessageEventListener{
 				event.setActive(false);
 			}else if(event.getMessage().getId() == GameCPM.DataPackage_EntityStatus) {
 				EntityStatusEvent entityStatus = gameCM.getClientPackageManager().readEntityStatusMessage(event.getMessage());
+				if(entityStatus == null) {
+					event.setActive(false);
+					return;
+				}
 				
 				entityStatus.getEntity().setLocation(entityStatus.getCurrentPixelLocation().getX(), entityStatus.getCurrentPixelLocation().getY());
 				if(!entityStatus.isAlive())	entityStatus.getEntity().destroy(false);	
 				event.setActive(false);
 			}else if(event.getMessage().getId() == GameCPM.DataPackage_ItemAdd) {
 				ItemAddEvent itemAdd = gameCM.getClientPackageManager().readItemAddMessage(event.getMessage());
+				if(itemAdd == null) {
+					return;
+				}
 				
 				itemAdd.getInv().addItemFunktion(itemAdd.getItem());
 				event.setActive(false);
 			}else if(event.getMessage().getId() == GameCPM.DataPackage_ItemRemove) {
 				ItemRemoveEvent itemRemove = gameCM.getClientPackageManager().readItemRemoveMessage(event.getMessage());
+				if(itemRemove == null) {
+					return;
+				}
 				
 				itemRemove.getInv().removeItemFunktion(itemRemove.getItem());
 				event.setActive(false);
 			}else if(event.getMessage().getId() == GameCPM.DataPackage_ItemSet) {
 				ItemSetEvent itemSet = gameCM.getClientPackageManager().readItemSetMessage(event.getMessage());
+				if(itemSet == null) {
+					return;
+				}
 				
 				itemSet.getInv().setItemFunktion(itemSet.getSlot(), itemSet.getItem());
 				event.setActive(false);
 			}else if(event.getMessage().getId() == GameCPM.DataPackage_MapBlockAdd) {
 				MapBlockAddEvent mapBlockAdd = gameCM.getClientPackageManager().readBlockAddMessage(event.getMessage());
+				if(mapBlockAdd == null) {
+					return;
+				}
 				
 				Map.getMap().add(mapBlockAdd.getResource().getID(), mapBlockAdd.getBlockLocation(), mapBlockAdd.getResource().isGround(), false);
 				event.setActive(false);
 			}else if(event.getMessage().getId() == GameCPM.DataPackage_MapBlockDelete) {
 				MapBlockDeleteEvent mapBlockDelete = gameCM.getClientPackageManager().readBlockDeleteMessage(event.getMessage());
+				if(mapBlockDelete == null) {
+					event.setActive(false);
+					return;
+				}
 				
 				Map.getMap().deleteBlock(mapBlockDelete.getMapBlock().getLocation(), mapBlockDelete.getMapBlock().getResource().getLayerUp(), mapBlockDelete.getMapBlock().getResource().isGround(), false);
+				event.setActive(false);
+			}else if(event.getMessage().getId() == GameCPM.DataPackage_DroneTarget) {
+				DroneTargetEvent droneTarget = gameCM.getClientPackageManager().readDroneTargetMessage(event.getMessage());
+				if(droneTarget == null) {
+					return;
+				}				
+				
+				DroneHost host = droneTarget.getHost();
+				if(droneTarget.isBuild()) {
+					host.addBuildTarget(droneTarget.getBlockLoc(), droneTarget.getResID());
+				} else {
+					host.addDestructionTarget(droneTarget.getBlockLoc(), droneTarget.isAdd());
+				}
 				event.setActive(false);
 			}
 		}catch (Exception e) {
